Login/requirements.txt
flask==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
itsdangerous==2.2.0
Werkzeug==3.0.3